Parse fields to send
